- Library 的 ConvertLineToPoints() 方法將部分 int 修改為 double - EventRecorder 將 VehicleName 中的 - 符號修改為單字 Dash 後再產出 Sql 指令 (Sqlite 的 Table 名稱不允許 - 符號)

+++ TrafficControlTest/Library/Library.cs
@@ -502,21 +502,21 @@ namespace TrafficControlTest.Library
 				double m = (double)(Start.mY - End.mY) / (Start.mX - End.mX);
 				double c = Start.mY - m * Start.mX;
 				double radian = Math.Atan((Start.mY - End.mY) / (Start.mX - End.mX));
-				int distance = Math.Abs((int)(Interval * Math.Cos(radian)));
+				double distance = Math.Abs((Interval * Math.Cos(radian)));
 				if (Start.mX < End.mX)
 				{
-					for (int i = Start.mX + distance; i < End.mX; i += distance)
+					for (double i = Start.mX + distance; i < End.mX; i += distance)
 					{
 						double y = Math.Round(m * i + c, 0, MidpointRounding.AwayFromZero);
-						result.Add(GenerateIPoint2D(i, (int)y));
+						result.Add(GenerateIPoint2D((int)i, (int)y));
 					}
 				}
 				else
 				{
-					for (int i = Start.mX - distance; i > End.mX; i -= distance)
+					for (double i = Start.mX - distance; i > End.mX; i -= distance)
 					{
 						double y = Math.Round(m * i + c, 0, MidpointRounding.AwayFromZero);
-						result.Add(GenerateIPoint2D(i, (int)y));
+						result.Add(GenerateIPoint2D((int)i, (int)y));
 					}
 				}
 			}
